Add Main component
The Main component will ensure the page at least covers the full
viewport height. This is done by stretching the main content of the page
(everything not in the header or the footer) with a min-height
attribute with respect to the viewport height and the heights of the
header and the footer.