Replace bootstrap.css with styled-normalize
Only a few parts of bootstrap were being used: reboot, card-link,
nav-link, and container/container-fluid. These pieces have already been
or will be readded to the project (most likely as styled-components) in
the neighboring commits.

This commit reimplements the style reset (as styled-normalize) and a few
random bits of global style that depend on the styled-components theme.

This is possible thanks to styled-components's `createGlobalStyle`
function. However, this method has a couple disadvantages due to
styled-components's inlined nature:

- The style reset and static global styles override external css
  like highlight.js's stylesheet.

- These styles, which are global throughout the site, are listed
  in full in every page's head. In contrast, if they were CSS files, they could be
  loaded once, cached, and loaded in a fraction of the time.

In the future, it would be wise to refactor the style reset and the
static global styles (default typography, spacing, etc.) into CSS files.
Track zeit/next.js#8626 for updates on better support for external CSS
in Next.js.